refactor: 4단계 MVC 패턴
