+++ src/main/java/domain/NameRedundancyException.java
@@ -1,2 +1,4 @@
+package domain;
+
 public class NameRedundancyException extends RuntimeException {
 }
