+++ src/main/java/domain/RandomGenerator.java
@@ -1,3 +1,5 @@
+package domain;
+
 import java.util.Random;
 
 public class RandomGenerator {
