+++ src/main/java/domain/WrongArrayLengthException.java
@@ -1,2 +1,4 @@
+package domain;
+
 public class WrongArrayLengthException extends RuntimeException {
 }
