+++ src/main/java/view/InputView.java
@@ -1,0 +1,21 @@
+package view;
+
+import java.util.Scanner;
+
+public class InputView {
+    private final Scanner scanner;
+
+    public InputView() {
+        this.scanner = new Scanner(System.in);
+    }
+
+    public String[] getCarNames() {
+        System.out.println("경주할 자동차 이름을 입력하세요(이름은 쉼표(,)를 기준으로 구분).");
+        return scanner.nextLine().split(",");
+    }
+
+    public int getMoveCount() {
+        System.out.println("시도할 횟수는 몇회인가요?");
+        return scanner.nextInt();
+    }
+}
