+++ src/test/java/domain/CarTest.java
@@ -1,3 +1,5 @@
+package domain;
+
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
